Project Complete!

+++ libft/ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: diomarti <diomarti@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/24 12:46:03 by diomarti          #+#    #+#             */
-/*   Updated: 2022/10/25 17:16:40 by diomarti         ###   ########.fr       */
+/*   Updated: 2022/11/09 15:13:57 by diomarti         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 size_t	ft_strlcpy(char *dst, const char *src, size_t size)
 {
 	size_t	i;
-	int		src_len;
+	size_t	src_len;
 
 	i = 0;
 	src_len = ft_strlen(src);
@@ -30,3 +30,5 @@ size_t	ft_strlcpy(char *dst, const char *src, size_t size)
 	}
 	return (src_len);
 }
+
+/*vai copiar size bytes do src para o dest*/
